requirements.txt
aiogram==3.21.0
pydantic-settings==2.10.1
SQLAlchemy==2.0.42
asyncpg==0.30.0
python-dateutil==2.9.0.post0